# Conflicts:
#	pom.xml
#	src/main/java/com/activity1/Activity1BackApplication.java
#	src/main/java/com/activity1/models/Categoria.java
#	src/main/java/com/activity1/models/Producto.java

+++ src/main/java/com/activity1/app/domain/repository/ProductosRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductosRepository extends JpaRepository<Producto, Integer> {
+public interface ProductosRepository extends JpaRepository<Producto, Long> {
 }
